+++ packages/react-sdk/src/hooks/usePersistedDevicePreferences.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import { CallingState } from '@stream-io/video-client';
+import { CallingState, InputDeviceStatus } from '@stream-io/video-client';
 import { useCallStateHooks } from '@stream-io/video-react-bindings';
 
 export type LocalDevicePreference = {
@@ -25,7 +25,7 @@ type DeviceState<K extends DeviceKey> = {
 };
 
 interface DeviceManagerLike {
-  state: { selectedDevice: string | undefined };
+  state: { selectedDevice: string | undefined; status?: InputDeviceStatus };
   select: (deviceId: string) => Promise<void> | void;
   enable?: () => Promise<void>;
   disable?: () => Promise<void>;
@@ -43,121 +43,228 @@ export const usePersistedDevicePreferences = (
 ): void => {
   const {
     useCallSettings,
-    useCameraState,
+    useCallCallingState,
     useMicrophoneState,
+    useCameraState,
     useSpeakerState,
   } = useCallStateHooks();
   const settings = useCallSettings();
-
-  usePersistedDevicePreference(
-    key,
-    'camera',
-    useCameraState(),
-    settings ? !settings.video.camera_default_on : undefined,
-  );
-
-  usePersistedDevicePreference(
-    key,
-    'microphone',
-    useMicrophoneState(),
-    settings ? !settings.audio.mic_default_on : undefined,
-  );
-
-  usePersistedDevicePreference(key, 'speaker', useSpeakerState(), false);
-};
-
-const usePersistedDevicePreference = <K extends DeviceKey>(
-  key: string,
-  deviceKey: K,
-  state: DeviceState<K>,
-  defaultMuted?: boolean,
-): void => {
-  const { useCallCallingState } = useCallStateHooks();
   const callingState = useCallCallingState();
+
+  const microphoneState: DeviceState<'microphone'> = useMicrophoneState();
+  const cameraState: DeviceState<'camera'> = useCameraState();
+  const speakerState: DeviceState<'speaker'> = useSpeakerState();
+
   const [applyingState, setApplyingState] = useState<
     'idle' | 'applying' | 'applied'
   >('idle');
-  const manager = state[deviceKey];
 
   useEffect(
     function apply() {
       if (
         callingState === CallingState.LEFT ||
-        !state.devices?.length ||
-        typeof defaultMuted !== 'boolean' ||
+        !microphoneState.devices.length ||
+        !cameraState.devices.length ||
+        !speakerState.devices ||
+        !settings ||
         applyingState !== 'idle'
       ) {
         return;
       }
 
-      const preferences = parseLocalDevicePreferences(key);
-      const preference = preferences[deviceKey];
-
       setApplyingState('applying');
 
-      if (!manager.state.selectedDevice) {
-        const applyPromise = preference
-          ? applyLocalDevicePreference(
-              manager,
-              [preference].flat(),
-              state.devices,
-            )
-          : applyMutedState(manager, defaultMuted);
+      (async () => {
+        for (const [deviceKey, state, defaultMuted] of [
+          ['microphone', microphoneState, !settings.audio.mic_default_on],
+          ['camera', cameraState, !settings.video.camera_default_on],
+          ['speaker', speakerState, false],
+        ] as const) {
+          const preferences = parseLocalDevicePreferences(key);
+          const preference = preferences[deviceKey];
+          const manager = (
+            state as DeviceState<'camera' | 'microphone' | 'speaker'>
+          )[deviceKey];
 
-        applyPromise
-          .catch((err) => {
+          const applyPromise = preference
+            ? applyLocalDevicePreference(
+                manager,
+                [preference].flat(),
+                state.devices,
+              )
+            : applyMutedState(manager, defaultMuted);
+
+          await applyPromise.catch((err) => {
             console.warn(
               `Failed to apply ${deviceKey} device preferences`,
               err,
             );
-          })
-          .finally(() => setApplyingState('applied'));
-      } else {
-        setApplyingState('applied');
-      }
+          });
+        }
+      })().finally(() =>
+        setApplyingState((state) => (state === 'applying' ? 'applied' : state)),
+      );
     },
     [
       applyingState,
       callingState,
-      defaultMuted,
-      deviceKey,
+      cameraState,
+      cameraState.devices,
       key,
-      manager,
-      state.devices,
+      microphoneState,
+      microphoneState.devices,
+      settings,
+      speakerState,
+      speakerState.devices,
     ],
   );
 
   useEffect(
     function persist() {
-      if (
-        callingState === CallingState.LEFT ||
-        !state.devices?.length ||
-        applyingState !== 'applied'
-      ) {
+      if (callingState === CallingState.LEFT || applyingState !== 'applied') {
         return;
       }
 
-      try {
-        patchLocalDevicePreference(key, deviceKey, {
-          devices: state.devices,
-          selectedDevice: state.selectedDevice,
-          isMute: state.isMute,
-        });
-      } catch (err) {
-        console.warn(`Failed to save ${deviceKey} device preferences`, err);
+      for (const [deviceKey, devices, selectedDevice, isMute] of [
+        [
+          'camera',
+          cameraState.devices,
+          cameraState.selectedDevice,
+          cameraState.isMute,
+        ],
+        [
+          'microphone',
+          microphoneState.devices,
+          microphoneState.selectedDevice,
+          microphoneState.isMute,
+        ],
+        [
+          'speaker',
+          speakerState.devices,
+          speakerState.selectedDevice,
+          speakerState.isMute,
+        ],
+      ] as const) {
+        try {
+          patchLocalDevicePreference(key, deviceKey, {
+            devices,
+            selectedDevice,
+            isMute,
+          });
+        } catch (err) {
+          console.warn(`Failed to save ${deviceKey} device preferences`, err);
+        }
       }
     },
     [
       applyingState,
       callingState,
-      deviceKey,
+      cameraState.devices,
+      cameraState.isMute,
+      cameraState.selectedDevice,
       key,
-      state.devices,
-      state.isMute,
-      state.selectedDevice,
+      microphoneState.devices,
+      microphoneState.isMute,
+      microphoneState.selectedDevice,
+      speakerState.devices,
+      speakerState.isMute,
+      speakerState.selectedDevice,
     ],
   );
 };
+
+// const usePersistedDevicePreference = <K extends DeviceKey>(
+//   key: string,
+//   deviceKey: K,
+//   state: DeviceState<K>,
+//   defaultMuted?: boolean,
+// ): void => {
+//   const { useCallCallingState } = useCallStateHooks();
+//   const callingState = useCallCallingState();
+//   const [applyingState, setApplyingState] = useState<
+//     'idle' | 'applying' | 'applied'
+//   >('idle');
+//   const manager = state[deviceKey];
+
+//   useEffect(
+//     function apply() {
+//       if (
+//         callingState === CallingState.LEFT ||
+//         !state.devices?.length ||
+//         typeof defaultMuted !== 'boolean' ||
+//         applyingState !== 'idle'
+//       ) {
+//         return;
+//       }
+
+//       const preferences = parseLocalDevicePreferences(key);
+//       const preference = preferences[deviceKey];
+
+//       setApplyingState('applying');
+
+//       if (!manager.state.selectedDevice) {
+//         const applyPromise = preference
+//           ? applyLocalDevicePreference(
+//               manager,
+//               [preference].flat(),
+//               state.devices,
+//             )
+//           : applyMutedState(manager, defaultMuted);
+
+//         applyPromise
+//           .catch((err) => {
+//             console.warn(
+//               `Failed to apply ${deviceKey} device preferences`,
+//               err,
+//             );
+//           })
+//           .finally(() => setApplyingState('applied'));
+//       } else {
+//         setApplyingState('applied');
+//       }
+//     },
+//     [
+//       applyingState,
+//       callingState,
+//       defaultMuted,
+//       deviceKey,
+//       key,
+//       manager,
+//       state.devices,
+//     ],
+//   );
+
+//   useEffect(
+//     function persist() {
+//       if (
+//         callingState === CallingState.LEFT ||
+//         !state.devices?.length ||
+//         applyingState !== 'applied'
+//       ) {
+//         return;
+//       }
+
+//       try {
+//         patchLocalDevicePreference(key, deviceKey, {
+//           devices: state.devices,
+//           selectedDevice: state.selectedDevice,
+//           isMute: state.isMute,
+//         });
+//       } catch (err) {
+//         console.warn(`Failed to save ${deviceKey} device preferences`, err);
+//       }
+//     },
+//     [
+//       applyingState,
+//       callingState,
+//       deviceKey,
+//       key,
+//       state.devices,
+//       state.isMute,
+//       state.selectedDevice,
+//     ],
+//   );
+// };
 
 const parseLocalDevicePreferences = (key: string): LocalDevicePreferences => {
   const preferencesStr = window.localStorage.getItem(key);
@@ -235,7 +342,10 @@ const applyLocalDevicePreference = async (
       devices.find((d) => d.label === p.selectedDeviceLabel);
 
     if (device) {
-      await manager.select(device.deviceId);
+      if (!manager.state.selectedDevice) {
+        await manager.select(device.deviceId);
+      }
+
       muted = p.muted;
       break;
     }
@@ -247,7 +357,9 @@ const applyLocalDevicePreference = async (
 };
 
 const applyMutedState = async (manager: DeviceManagerLike, muted: boolean) => {
-  await manager[muted ? 'disable' : 'enable']?.();
+  if (!manager.state.status) {
+    await manager[muted ? 'disable' : 'enable']?.();
+  }
 };
 
 const getSelectedDevicePreference = (